Change Delete

+++ src/Datiss.Budget.Services/WasteSalesSplitService.cs
@@ -213,13 +213,21 @@ namespace Datiss.Budget.Services
             var self = await _dbSet.Where(_ => _.YearId == yearId)
                                     .Where(_ => _.OrganizationId == organizationId)
                                     .ToListAsync();
-            _dbSet.RemoveRange(self);
 
-            var childrens = await getChildren(organizationId, yearId);
-            _dbSet.RemoveRange(childrens);
+            IEnumerable<WasteSalesSplit> childrens = new WasteSalesSplit[] { };
+
+            if (organization.Type == OrganizationType.County || organization.Type == OrganizationType.Root)
+            {
+                childrens = await getChildren(organizationId, yearId);
+            }
 
             if (self.Count() == 0 && childrens.Count() == 0)
                 throw new DeleteNullRecordException();
+
+            _dbSet.RemoveRange(self);
+
+            if (childrens.Any())
+                _dbSet.RemoveRange(childrens);
 
             var result = new OrganizationDeleteDataResult
             {
